Update 0.1
I changed the way temperature is displayed. Now it is only average because if I put it like that (min-max), than there is a high chance the bot will put two identical numbers, although they are float.

Another thing — now, instead of telling what's going on in the skies (clear sky, rain, etc.) it outputs corresponding emoji. It's still should be adjusted according to the time in the place because if there is night and the sky is clear, my bot will output the Sun emoji.